feat: convert new-app-review to typescript [EXT-5577] (#2520)
