+++ .github/workflows/new-app-review/validators/dependencies.ts
@@ -1,0 +1,48 @@
+import { spawn } from 'child_process';
+import type { ValidatorOptions, ValidatorResult, PullRequestFile } from '../../types';
+
+export const validate = async (_options: ValidatorOptions, newAppDir: string, _files: PullRequestFile[]): Promise<ValidatorResult> => {
+  return new Promise((resolve, reject) => {
+    const command = `npm outdated --json`;
+    const child = spawn(command, { shell: true, cwd: newAppDir });
+
+    let output = '';
+    let errorOutput = '';
+
+    child.stdout.on('data', (data) => {
+      output += data.toString();
+    });
+
+    child.stderr.on('data', (data) => {
+      errorOutput += data.toString();
+    });
+
+    child.on('close', (code) => {
+      if (code === 0) {
+        resolve({
+          result: true,
+          message: 'Dependency check finished',
+        });
+      } else if (output) {
+        let warning = '';
+        const outdatedDependencies = JSON.parse(output);
+        if (Object.keys(outdatedDependencies).length > 0) {
+          warning += 'The following dependencies are outdated:\n';
+          warning += Object.keys(outdatedDependencies)
+            .map((dependency) => `- ${dependency}`)
+            .join('\n');
+        }
+        resolve({
+          result: true,
+          warning,
+          message: 'Dependency check finished',
+        });
+      } else {
+        resolve({
+          result: false,
+          message: 'Failed to check for outdated dependencies',
+        });
+      }
+    });
+  });
+};
