+++ package.json
@@ -7,6 +7,9 @@
   },
   "packageManager": "npm@8.19.4",
   "devDependencies": {
+    "@actions/core": "^1.10.1",
+    "@actions/github": "^6.0.0",
+    "@types/node": "^22.3.0",
     "husky": "^8.0.0",
     "lint-staged": "^13.2.1",
     "prettier": "2.8.8"
